Add PAR parameter selection and residual printing
Some details has been changed as well such as the order of the
parameters taken by run_mumps.sh. It now follow a much common way of
number of MPI ranks then number of threads.

Some changes has been applied to the mlkaps configuration and launching
script to follow the current state of the experiments.