Change linking to multithreaded BLAS/LAPACK
The previous build was linked with the sequential MKL.
Therefore, we observed no scaling while increasing the number of OMP
threads. The link flags has been changed to use the multithreaded
version of the MKL.

Furthermore, there were issues with SLURM allocation which force us to
use a workaround to launch jobs with the exact allocation.