Refactor: remove resolving stage and change names
The third stage of the mumps resolution has been deactivated as we are
mostly interested in the performance of the factorization (stage 2).
As a consequence, we don't need a rhs anymore. The code associated with
it hasn't been removed as we plan to reactivate the third stage through a
CLI option.

Datatype of the matrix elements has been renamed to be easier to
understand.