Add density and bandwidth parameters
The user no longer choose the nnz of the matrix but rather the density
of the band. The nnz is computed thanks to the n and bandwidth
parameters.

The clang-format has been repaired. It resulted in some minor changes of
style in existing files.

Printings of the generated matrix put during the debugging were still
present. They have been removed.