Add Energy measurement
The energy measurements are done through EAR API.
As the installation process of EAR is complex, this dependency isn't
mandatory.
A conditional compiled is performed to activate EAR whether the library
is found or not.
If anyone wants to have the energy info he has to give the `-Dear-path`
option to `meson setup`.

Some minor improvements have been made to the compilation steps.
A better specialization between the compilers has been performed.