Rework launching scripts
`run_mumps.sh` is now a generic script to launch any type of MUMPS
experiment we thought of.
Auxiliary scripts have been added to provide a simple target for
applications (i.e. mlkaps).
All the scripts are installed to the `bin` directory with the built
binary during the installation phase of a meson run.